refactor : 폴더 구조 정리 - 네비(#98) - 네비바 폴더 구조 세분화, 분리
