+++ client/src/constant/index.js
@@ -1,0 +1,5 @@
+export const ProfileInfo = Object.freeze({
+  BLOG: "https://wonderfulharu.tistory.com/",
+  GITHUB: "https://github.com/psy-g",
+  EMAIL: "mailto:psykyg@gmail.com",
+});
